Congress Votes Project
In this project I used a database of congress votes.
This is the link to the database and the explanation about it:
http://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records

In this project, I built the Entropy for the entire congress' votes, and
calculated the information-gain per each attribute.
I then took the maximum info-gain, and constructed a one-level decision
tree with an accuracy of 90%, according to 10 random rows picked.